web_compliance_system/requirements.txt
# Spes合规审查系统 - Web版本依赖

# Web框架
Flask==2.3.3
Werkzeug==2.3.7

# LangChain相关
langchain==0.3.27
langchain-openai==0.3.33
langchain-community==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11

# 向量数据库
faiss-cpu==1.8.0

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8

# 图片处理
Pillow==11.3.0
opencv-python==4.10.0.84

# HTTP请求
requests==2.32.4
httpx==0.28.1

# 环境变量
python-dotenv==1.1.1

# 其他工具
numpy==1.26.4
pydantic==2.11.9
tiktoken==0.11.0
